requirements.txt
cantools
python_can